scripts/data_pipeline/requirements.txt
langchain
langchain-openai
langchain-community
requests
beautifulsoup4
pinecone-client
python-dotenv
openai
tiktoken
